LayoutTests/http/tests/loading/basic-auth-resend-wrong-credentials-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/test2/protected-resource.php - didReceiveAuthenticationChallenge - Responding with wrongusername:wrongpassword
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/test2/basic-auth-testing.php?username=correctusername&password=correctpassword - didReceiveAuthenticationChallenge - Responding with correctusername:correctpassword
frame "<!--framePath //<!--frame1-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishLoadForFrame
This test makes sure that once WebCore preemptively sends out Basic credentials it thinks apply to a new resource, and that resource response with a 401 challenge, that it doesn't try to send the same wrong credentials a second time.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Authenticated as user: wrongusername password: wrongpassword

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Authenticated as user: correctusername password: correctpassword
